refactor(sparq-solid): extract WAC/ACP parity corpus to a reusable test source (sq-t58w.6)
[OPUS-4.8] Pure test-infra refactor enabling the upcoming differential oracle
(sq-t58w.7) per research/solid-acp-differential-oracle-design.md §3.2/§5.

The WAC/ACP scenario corpus (the AclBuilder/AcrBuilder-emitted .acl/.acr data +
the Expect decision tables, 12 WAC + 12 ACP) lived inline in conformance_wac.rs
/ conformance_acp.rs. It is moved verbatim into a shared tests/common module —
common::wac_corpus() / common::acp_corpus() — so a SECOND test target can
consume the IDENTICAL scenarios without copy-paste. "One corpus, two consumers."

No behaviour change: the scenario builder bodies and corpus orderings are
byte-identical to the originals (verified by diff); the two conformance files
remain the corpus's first consumers with the same test set + assertions. Test
counts unchanged: conformance_wac = 3 tests, conformance_acp = 2 tests, in both
the default and odrl-bridge feature states.

Docs: skills/access-control/SKILL.md re-points the corpus-location notes to the
single reusable source.
